+++ packages/openapi-ts/src/config/plugins.ts
@@ -4,8 +4,8 @@ import type {
   PluginContext,
   PluginNames,
 } from '~/plugins/types';
-import type { Config, UserConfig } from '~/types/config';
 
+import type { Config, UserConfig } from './types';
 import { valueToObject } from './utils/config';
 import { packageFactory } from './utils/package';