+++ packages/types/CHANGELOG.md
@@ -1,5 +1,11 @@
 # @hey-api/types
 
+## 0.1.2
+
+### Patch Changes
+
+- feat: add `ToArray`, `ToReadonlyArray`, and `AnyObject` types ([#3244](https://github.com/hey-api/openapi-ts/pull/3244)) ([`4f52bce`](https://github.com/hey-api/openapi-ts/commit/4f52bce79d8ac6573472f32a05b7f70913b90605)) by [@mrlubos](https://github.com/mrlubos)
+
 ## 0.1.1
 
 ### Patch Changes